unfortunate move+refactor of picsv batch stuff
moves most of the CSV logic out of cmd/picsv and into csv
package. refactor to be a bit cleaner and get ready to handle multiple
sources of CSV data.
